+++ linux_c_develop/mydup2.c
@@ -38,7 +38,7 @@ int main9(int argc, char* argv1[])
 	//使用newfd读文件
 	char buf[64];
 	memset(buf, 0x00, sizeof(buf));
-	int n = read(oldfd, buf, sizeof(buf));
+	ssize_t n = read(oldfd, buf, sizeof(buf));
 	printf("read over: n==[%d], buf==[%s]\n", n, buf);
 
 	//关闭文件
